+++ src/constants.ts
@@ -4,6 +4,7 @@
 
 export const RESPONSE_ERROR_MESSAGE = {
     500: 'Failed to validate key.',
+    501: 'You do not have persmision to do this.',
     400: 'Looks like an error occurred. Try again later.',
     403: 'Inccorrect Use of Params.',
     404: 'No Signature Header Was Provided.',
