+++ src/controller/utils/CVE_Object_Class.ts
@@ -9,7 +9,6 @@ export class CveObject {
     public CVE_ID_OBJECT = {
         cve_id: '',
         status: '',
-        Requestor: { user_id: '', user_name: '' },
         Organization: { org_id: '', org_name: '' },
         History: 'TBD',
     };
@@ -31,13 +30,10 @@ export class CveObject {
     }
 
     public createCVEID() {
-        this.CVE_ID_OBJECT.cve_id = 'CVE-' + this.year + '-' + uuidv4();
+        this.CVE_ID_OBJECT.cve_id = 'CVE-' + this.year + '-' + uuidv4(); //can short the uuid to last 4 digits if needed
         this.CVE_ID_OBJECT.status = this.status;
-        this.CVE_ID_OBJECT.Requestor.user_id = this.request.userData.userUUID;
-        this.CVE_ID_OBJECT.Requestor.user_name =
-            this.request.userData.userFName + ' ' + this.request.userData.userLName;
-        this.CVE_ID_OBJECT.Organization.org_id = this.request.userData.userCNAUUID;
-        this.CVE_ID_OBJECT.Organization.org_name = this.request.userData.userCNAName;
+        this.CVE_ID_OBJECT.Organization.org_id = this.request.cnaData.cnaUUID;
+        this.CVE_ID_OBJECT.Organization.org_name = this.request.cnaData.cnaName;
         return this.CVE_ID_OBJECT;
     }
 }
