+++ src/database/db-operations.ts
@@ -2,7 +2,7 @@ import * as _ from 'lodash';
 
 const db = require('./models/index'); // tslint:disable
 import { CNADetailsObject } from './dbconstants';
-import { IUsers } from './interfaces/db-interfaces';
+import { ICNA } from './interfaces/db-interfaces';
 
 export const dbOperations = {
     /**
@@ -85,55 +85,45 @@ export const dbOperations = {
      * @returns {Object} - Returns users data pertaining to the UUID
      */
     async getUserData(UUID) {
-        return db.CNAs.findOne({
+        return db.Admins.findOne({
             where: {
-                cnaUUID: UUID,
-                cnaStatus: 'ACTIVE',
-            },
+                adminUUID: UUID
+            }
         })
             .then(
                 resolved => {
-                    const { cnaPublicKey, cnaUUID, cnaName } = resolved;
-                    console.log({ cnaPublicKey, cnaUUID, cnaName });
-                    return Promise.resolve({ cnaPublicKey, cnaUUID, cnaName });
+                    const { adminPublicKey, adminUUID } = resolved;
+                    return Promise.resolve({ adminPublicKey, adminUUID });
                 },
                 rejection => {},
             )
             .catch(err => {
-                return db.Users.findOne({
+                return db.CNAs.findOne({
                     where: {
-                        userUUID: UUID,
-                        userStatus: 'ACTIVE',
+                        cnaUUID: UUID,
+                        cnaStatus: 'ACTIVE',
                     },
                 }).then(
                     resolved => {
                         const {
-                            userPublicKey,
-                            userUUID,
-                            userType,
-                            userHardQuota,
-                            userSoftQuota,
-                            userDescription,
-                            userGithubID,
-                            userFName,
-                            userLName,
-                            userEmail,
-                            userCNAUUID,
-                            userCNAName,
+                            cnaUUID,
+                            cnaPublicKey,
+                            cnaName,
+                            cnaType,
+                            cnaDescription,
+                            cnaEmail,
+                            cnaWebSite,
+                            cnaGPGKeys
                         } = resolved;
                         return Promise.resolve({
-                            userPublicKey,
-                            userUUID,
-                            userType,
-                            userHardQuota,
-                            userSoftQuota,
-                            userDescription,
-                            userGithubID,
-                            userFName,
-                            userLName,
-                            userEmail,
-                            userCNAUUID,
-                            userCNAName,
+                            cnaUUID,
+                            cnaPublicKey,
+                            cnaName,
+                            cnaType,
+                            cnaDescription,
+                            cnaEmail,
+                            cnaWebSite,
+                            cnaGPGKeys
                         });
                     },
                     rejection => {
@@ -151,8 +141,6 @@ export const dbOperations = {
         var data = {
             cveID: cveObject.cve_id,
             cveStatus: cveObject.status,
-            cveRequesterUserID: cveObject.Requestor.user_id,
-            cveRequesterUserName: cveObject.Requestor.user_name,
             cveOrganizationUUID: cveObject.Organization.org_id,
             cveOragnizationName: cveObject.Organization.org_name,
         };
@@ -188,9 +176,11 @@ export const dbOperations = {
         )
             .then(
                 isInDb => {
+                    console.log(isInDb[0])
                     return Promise.resolve(isInDb);
                 },
                 rejection => {
+                    console.log("here2")
                     return Promise.reject(rejection);
                 },
             )
@@ -204,11 +194,55 @@ export const dbOperations = {
      * @param {string} userStatus - Status to give to the User
      * @param {string} userCNAUUID - Organization pertaining to the User
      */
-    async deactivateUser(userUUID: string, userCNAUUID: string) {
-        return db.Users.update(
-            { userStatus: 'INACTIVE' },
+    async deactivateCNA(cnaUUID: string) {
+        return db.CNAs.update(
+            { cnaStatus : 'INACTIVE' },
             {
-                where: { userUUID, userCNAUUID },
+                where: {cnaUUID}
+            },
+        )
+            .then(
+                resolved => {
+                    console.log(resolved)
+                    return Promise.resolve(resolved);
+                },
+                rejection => {
+                    return Promise.reject(rejection);
+                },
+            )
+            .catch(err => {
+                console.log(err)
+                return Promise.reject({ error: err });
+            });
+    },
+    /**
+     * Edits the status of a given User ID
+     * @param {string} cnaUUID - User ID to edit status
+     * @param {Object} user - User Object
+     * @param {string} userCNAUUID - Organization pertaining to the User making the call
+     */
+    async updateCNA(cnaUUID: string, cna: ICNA) {
+        const {
+            cnaPublicKey,
+            cnaName,
+            cnaType,
+            cnaDescription,
+            cnaEmail,
+            cnaWebSite,
+            cnaGPGKeys
+        } = cna;
+        return db.CNAs.update(
+            {
+                cnaPublicKey,
+            cnaName,
+            cnaType,
+            cnaDescription,
+            cnaEmail,
+            cnaWebSite,
+            cnaGPGKeys
+            },
+            {
+                where: { cnaUUID },
             },
         )
             .then(
@@ -223,50 +257,21 @@ export const dbOperations = {
                 return Promise.reject({ error: err });
             });
     },
+
     /**
-     * Edits the status of a given User ID
-     * @param {string} userUUID - User ID to edit status
-     * @param {Object} user - User Object
-     * @param {string} userCNAUUID - Organization pertaining to the User making the call
+     * Create a CNA with a given cna object
+     * @param {CNA} cna - User ID to edit status
+     * @returns {string} - Returns the new CNA UUID
      */
-    async updateUser(userUUID: string, userCNAUUID: string, user: IUsers) {
-        const {
-            userType,
-            userGithubID,
-            userDescription,
-            userFName,
-            userLName,
-            userEmail,
-            userURL,
-            userGPGKeys,
-            userPublicKey,
-        } = user;
-        return db.Users.update(
-            {
-                userType,
-                userGithubID,
-                userDescription,
-                userFName,
-                userLName,
-                userEmail,
-                userURL,
-                userGPGKeys,
-                userPublicKey,
+    async createCNA(cna) {
+        return db.CNAs.create(cna).then(
+            resolve => {
+                const { cnaUUID } = resolve;
+                return Promise.resolve(cnaUUID);
             },
-            {
-                where: { userUUID, userCNAUUID },
+            rejected => {
+                return Promise.reject(rejected);
             },
-        )
-            .then(
-                resolved => {
-                    return Promise.resolve(resolved);
-                },
-                rejection => {
-                    return Promise.reject(rejection);
-                },
-            )
-            .catch(err => {
-                return Promise.reject({ error: err });
-            });
+        );
     },
 };
