+++ src/database/interfaces/db-interfaces.ts
@@ -1,11 +1,9 @@
-export interface IUsers {
-    userPublicKey?: string;
-    userType?: string;
-    userDescription?: string;
-    userGithubID: string;
-    userFName?: string;
-    userLName?: string;
-    userEmail?: string;
-    userURL?: string;
-    userGPGKeys?: string
+export interface ICNA {
+    cnaPublicKey?: string;
+    cnaName?: string;
+    cnaType?: string;
+    cnaDescription?: string;
+    cnaEmail?: string;
+    cnaWebSite?: string;
+    cnaGPGKeys?: string
   }
