+++ src/database/models/admin.js
@@ -1,0 +1,43 @@
+"use strict";
+module.exports = (sequelize, DataTypes) => {
+  var Admins = sequelize.define(
+    "Admins",
+    {
+      adminUUID: {
+        type: DataTypes.UUID,
+        defaultValue: DataTypes.UUIDV4,
+        allowNull: false,
+        primaryKey: true
+      },
+      adminPublicKey: {
+        type: DataTypes.TEXT("long"),
+        allowNull: false
+      },
+      adminEmail: {
+        type: DataTypes.STRING,
+        validate: {
+          isEmail: true
+        },
+        unique: true
+      },
+      createdAt: {
+        allowNull: false,
+        type: DataTypes.DATE
+      },
+      updatedAt: {
+        allowNull: false,
+        type: DataTypes.DATE
+      },
+      deletedAt: {
+        allowNull: true,
+        type: DataTypes.DATE
+      }
+    },
+    {
+      timestamps: true,
+      paranoid: true
+    }
+  );
+  Admins.associate = function(models) {};
+  return Admins;
+};
