+++ src/database/models/cves.js
@@ -41,13 +41,6 @@ module.exports = (sequelize, DataTypes) => {
       cveOragnizationName: {
         type: DataTypes.STRING
       },
-      cveUserUUID: {
-        type: DataTypes.UUID,
-        references: {
-          model: "Users",
-          key: "userUUID"
-        }
-      },
       createdAt: {
         allowNull: false,
         type: DataTypes.DATE
