--- src/database/models/users.js
@@ -1,95 +1,0 @@
-"use strict";
-module.exports = (sequelize, DataTypes) => {
-  var Users = sequelize.define(
-    "Users",
-    {
-      userUUID: {
-        type: DataTypes.UUID,
-        defaultValue: DataTypes.UUIDV4,
-        allowNull: false,
-        primaryKey: true
-      },
-      userPublicKey: {
-        type: DataTypes.TEXT("long"),
-        allowNull: false
-      },
-      userStatus: {
-        type: DataTypes.ENUM,
-        values: ["ACTIVE", "INACTIVE"],
-        defaultValue: "ACTIVE"
-      },
-      userType: {
-        type: DataTypes.ENUM,
-        values: ["VENDOR", "RESEARCHER", "PUBLISHER"],
-        defaultValue: "VENDOR"
-      },
-      userHardQuota: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 100
-      },
-      userSoftQuota: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        defaultValue: 50
-      },
-      userDescription: {
-        type: DataTypes.STRING,
-        allowNull: false
-      },
-      userGithubID: {
-        type: DataTypes.STRING,
-        allowNull: false
-      },
-      userFName: {
-        type: DataTypes.STRING,
-        allowNull: false
-      },
-      userLName: {
-        type: DataTypes.STRING,
-        allowNull: false
-      },
-      userEmail: {
-        type: DataTypes.STRING,
-        validate: {
-          isEmail: true
-        },
-        unique: true
-      },
-      userURL: {
-        type: DataTypes.STRING
-      },
-      userGPGKeys: {
-        type: DataTypes.STRING
-      },
-      userCNAUUID: {
-        type: DataTypes.UUID,
-        references: {
-          model: "CNAs",
-          key: "cnaUUID"
-        }
-      },
-      userCNAName: {
-        type: DataTypes.STRING
-      },
-      createdAt: {
-        allowNull: false,
-        type: DataTypes.DATE
-      },
-      updatedAt: {
-        allowNull: false,
-        type: DataTypes.DATE
-      },
-      deletedAt: {
-        allowNull: true,
-        type: DataTypes.DATE
-      }
-    },
-    {
-      timestamps: true,
-      paranoid: true
-    }
-  );
-  Users.associate = function(models) {};
-  return Users;
-};
