+++ src/database/seeders/demo-admin.js
@@ -1,0 +1,23 @@
+"use strict";
+var {TEST_CNA_UUID, TEST_USER_PUBLICKEY, TEST_USER_UUID, TEST_CNA_COMPANY_NAME} = require('../../../test/constants');
+module.exports = {
+  up: (queryInterface, Sequelize) => {
+    return queryInterface.bulkInsert(
+      "Admins",
+      [
+        {
+          adminUUID: TEST_USER_UUID,
+          adminPublicKey : TEST_USER_PUBLICKEY,
+          adminEmail: 'admin@admin.com',
+          createdAt: new Date(),
+          updatedAt: new Date()
+        }
+      ],
+      {}
+    );
+  },
+
+  down: (queryInterface, Sequelize) => {
+    return queryInterface.bulkDelete("Admins", null, {});
+  }
+};
