+++ src/database/seeders/demo-cve.js
@@ -9,8 +9,6 @@ module.exports = {
           cveUUID: TEST_CVE_UUID,
           cveID : TEST_CVE_ID,
           cveStatus: TEST_CVE_STATUS,
-          cveRequesterUserID: TEST_USER_UUID,
-          cveRequesterUserName: "test description",
           cveOrganizationUUID: TEST_CNA_UUID,
           cveOragnizationName: TEST_CNA_COMPANY_NAME,
           createdAt: new Date(),
