--- src/database/seeders/demo-user.js
@@ -1,34 +1,0 @@
-"use strict";
-var {TEST_CNA_UUID, TEST_USER_PUBLICKEY, TEST_USER_UUID, TEST_CNA_COMPANY_NAME} = require('../../../test/constants');
-module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert(
-      "Users",
-      [
-        {
-          userUUID: TEST_USER_UUID,
-          userCNAUUID : TEST_CNA_UUID,
-          userFName: "Test user",
-          userLName: "Test lastname",
-          userDescription: "test description",
-          userType: "VENDOR",
-          userEmail: "nestor@teddy.com",
-          userURL: "urltest",
-          userHardQuota: 100,
-          userSoftQuota: 50,
-          userGithubID: 'tester',
-          userGPGKeys: 'rrrr',
-          userCNAName: TEST_CNA_COMPANY_NAME,
-          userPublicKey: TEST_USER_PUBLICKEY,
-          createdAt: new Date(),
-          updatedAt: new Date()
-        }
-      ],
-      {}
-    );
-  },
-
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.bulkDelete("Users", null, {});
-  }
-};
