+++ src/routes.ts
@@ -1,28 +1,28 @@
-import { CveIdAllocationController } from './controller/CveIdAllocationService';
-import { CNA } from './controller/CNAController';
+import { CNAController } from './controller/CNAController';
+import { Admin } from './controller/AdminController';
 import { DebugController } from './controller/DebugController';
 import { mapValues } from 'lodash';
-let cna = new CNA();
+let admin = new Admin();
 let debug = new DebugController();
-let cve = new CveIdAllocationController();
+let CNA = new CNAController();
 export const ROUTES = {
-    '/register-user': {
-        post: cna.registerUser,
+    '/register-cna': {
+        post: admin.registerCNA,
     },
-    '/deactivate-user': {
-        post: cna.deactivateUser,
+    '/deactivate-cna': {
+        post: admin.deactivateCNA,
     },
-    '/update-user': {
-        post: cna.updateUser,
+    '/update-cna': {
+        post: admin.updateCNA,
     },
     '/get-cve-id': {
-        post: cve.getCveId,
+        post: CNA.getCveId,
     },
     '/get-cvs-info': {
-        post: cve.getCveInfo,
+        post: CNA.getCveInfo,
     },
     '/set-cve-status': {
-        post: cve.setCveIdStatus,
+        post: CNA.setCveIdStatus,
     },
     '/health-check': {
         get: debug.health,
